Refactor: js파일들 경로 및 분류 수정
index.js에 중구난방으로 연결되어 있었던 여러 js파일들을 Header, Inputbar,
Main 영역 세 분류로 폴더를 나누어 각 기능을 담당하는 위치로 이동시키고 모듈화했다.
