+++ .eslintrc.json
@@ -13,7 +13,8 @@
     "no-param-reassign": 0,
     "import/prefer-default-export": 0,
     "import/extensions": 0,
-    "no-restricted-syntax": 0
+    "no-restricted-syntax": 0,
+    "no-use-before-define": 0,
   },
   "env": {
     "browser": true,
