+++ src/js/header/initHeader.js
@@ -1,0 +1,5 @@
+import { initHeaderMonthYear } from './initHeaderMonthYear.js';
+
+export function initHeader() {
+	initHeaderMonthYear();
+}
