--- src/js/inputbar/autoComma.js
@@ -1,6 +1,0 @@
-const inputPrice = document.querySelector('#inputPrice');
-
-inputPrice.addEventListener('input', e => {
-	e.target.value = e.target.value.replace(/,/g, '');
-	e.target.value = e.target.value.replace(/\B(?=(\d{3})+(?!\d))/g, ',');
-});
