+++ src/js/inputbar/initInputBarDate.js
@@ -1,6 +1,6 @@
 import { getToday } from '../utils/getToday.js';
 
-export const initInputBarDate = () => {
+export const initInputbarDate = () => {
 	const dateInput = document.querySelector('#inputDate');
 	let { year, month, date } = getToday();
 	month = `0${month}`.slice(-2);
