+++ src/js/main/initMain.js
@@ -1,0 +1,7 @@
+import { initMainRender } from './initMainRender.js';
+import { initFilterList } from './infoFilter.js';
+
+export function initMain() {
+	initMainRender();
+	initFilterList();
+}
