+++ src/js/utils/autoComma.js
@@ -1,0 +1,8 @@
+export function addAutoComma() {
+	const inputPrice = document.querySelector('#inputPrice');
+
+	inputPrice.addEventListener('input', e => {
+		e.target.value = e.target.value.replace(/,/g, '');
+		e.target.value = e.target.value.replace(/\B(?=(\d{3})+(?!\d))/g, ',');
+	});
+}
